Split new functionality into new command
